Import the revamped menu from ReGBA
While that menu revamp was actually mandatory in ReGBA due to gcc 5.2.0
emitting bad code for the previous version, here, it was not mandatory,
but it does look better.

DefaultDisplayValue now displays option values right-aligned. Spaces as
alignment are removed, as well as format codes, in language.msg.

The screen is no longer copied before drawing a message box.